Multi-Agent Deep Q Learning
Trains two models, a ground robot and a scouting drone. It simulates the robot being blind to obstacles, encouraging the drone to do the majority of the exploring by penalizing movements harder when simulated uncertainty is higher. Actions which would promote uncertainty in real life will punish the system harder. The drone is encouraged to explore the environment quickly and return to the ground robot, and the ground robot uses this information to stay close to the obstacles without hitting them. It is also encouraged to get to the goal as quickly as possible, as the punishment for waiting is exponentially increasing. Hopefully, the Network is able to find the optimal combination of waiting for the environment to be explored, and moving toward the goal in a quick but certain way.